+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Next]
 - Fix trace bug when all children disappear when the root children has over 20 items
+- Ensure consistent background on CM6
+- Show node details on snapshots
 
 ## [2021.04.22]
 - In the request tab, recent requests now show on top of the list